% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_terminal_ranked}
\alias{calc_terminal_ranked}
\title{Finds supersets for ranked permutations}
\usage{
calc_terminal_ranked(ranked_perm, current_range, target, i, n.filtered)
}
\arguments{
\item{ranked_perm}{data frame of credible hierarchies}

\item{current_range}{range of the target}

\item{target}{permutation we are looking for}

\item{i}{current index}

\item{n.filtered}{number of rows}
}
\description{
Finds supersets for ranked permutations
}
\keyword{internal}
