% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_terminal_unranked}
\alias{calc_terminal_unranked}
\title{Finds supersets for permutations}
\usage{
calc_terminal_unranked(perm, target, i, n.filtered)
}
\arguments{
\item{perm}{data frame of credible hierarchies}

\item{target}{permutation we are looking for}

\item{i}{current index}

\item{n.filtered}{number of rows}
}
\description{
Finds supersets for permutations
}
\keyword{internal}
