% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{compare_nonconsec}
\alias{compare_nonconsec}
\title{Compare posets with permutations}
\usage{
compare_nonconsec(df, target)
}
\arguments{
\item{df}{data frame of credible hierarchies}

\item{target}{poset we are looking for}
}
\description{
Compare posets with permutations
}
\keyword{internal}
