% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive.R
\name{consecutive}
\alias{consecutive}
\title{Retrieves credible hierarchies for consecutive question types}
\usage{
consecutive(hierarchy.matrix, threshold)
}
\arguments{
\item{hierarchy.matrix}{A matrix where column headers are ranks and each row displays the treatments assigned to each rank for that iteration}

\item{threshold}{A number that determines which hierarchies are credible}
}
\value{
A list of data frames containing the credible hierarchies for Ranked Permutations, Permutations, Ranked Combinations, and Combinations
}
\description{
Finds all Ranked Permutations, Permutations, Ranked Combinations, and Combinations with frequencies greater than or equal to the threshold
}
\examples{
consecutive(hierarchy.matrix,0.5)
}
