% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{create_perm}
\alias{create_perm}
\title{Creates permutations for poset function}
\usage{
create_perm(trts, trt1, new_trts, perm_size, credible)
}
\arguments{
\item{trts}{list of treatment names}

\item{trt1}{first treatment in the new pair}

\item{new_trts}{potential new treatments to add}

\item{perm_size}{size of the permutation}

\item{credible}{list of credible size two permutations}
}
\description{
Creates permutations for poset function
}
\keyword{internal}
