% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_Thijs2008.R
\docType{data}
\name{dat_Thijs2008}
\alias{dat_Thijs2008}
\title{Antiplatelet regimens for the prevention of serious vascular events after
transient ischaemic attack (TIA) or stroke}
\format{
A data frame with 15,000 rows and 6 columns:
\describe{
  \item{Chain}{The index of the chain.}
  \item{Placebo}{The sampled relative effects of placebo vs. placebo
  (should be 0).}
  \item{ASA}{The sampled relative effects of aspirin (ASA) vs. placebo.}
  \item{ASADP}{The sampled relative effects of aspirin + dipyridamol
  (ASA+DP) vs. placebo.}
  \item{Thieno}{The sampled relative effects of thienopyridines (Thieno)
  vs. placebo.}
  \item{ThienoASA}{The sampled relative effects of thienopyridines +
  aspirin (Thieno+ASA) vs. placebo.}
}
}
\source{
Thijs V, Lemmens R, Fieuws S. (2008). Network meta-analysis: simultaneous meta-analysis of common antiplatelet regimens after transient ischaemic attack or stroke, Eur Heart J 29(9):1086-92. [doi:10.1093/eurheartj/ehn106]. Adopted from nmadb::readByID(501404)$data.

data(dat_Thijs2008)
}
\usage{
dat_Thijs2008
}
\description{
Samples from the joint distribution of the relative effects for each
active treatment (aspirin, aspirin + dipyridamole, thienopyridines,
thienopyridines + aspirin) vs. placebo estimated in a fixed effect Bayesian
network meta-analysis with a binomial likelihood and logit link. Three chains
were run, the first 50,000 samples were discarded as burn-in, and the
subsequent 5,000 samples were thinned by 10, resulting in a total of 1,500
samples saved in this data frame. The outcome is serious vascular events.
}
\keyword{datasets}
