% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_redundancies.R
\name{find_redundancies}
\alias{find_redundancies}
\title{Identify and optionally trim redundant hierarchies}
\usage{
find_redundancies(
  algo_1,
  algo_2,
  algo_3,
  n_trt,
  threshold,
  type = 1:18,
  trim_redundant = FALSE
)
}
\arguments{
\item{algo_1}{a list of data frames outputted from `get_arrangements()`
containing the credible hierarchies for ranked permutations, permutations,
ranked combinations, and combinations.}

\item{algo_2}{a data frame outputted from `get_partial_hierarchies()`
containing credible partial hierarchies.}

\item{algo_3}{a data frame outputted from `get_ranks_by_treatments()`
containing credible HDRs.}

\item{n_trt}{a numeric value indicating the number of treatments in the
network.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}

\item{type}{a numeric vector indicating what types of hierarchy redundancies
should be trimmed. See details for more information. Default is 1:18.}

\item{trim_redundant}{a logical value indicating whether the redundant
hierarchies should be trimmed from the output (TRUE) or not (FALSE, the
default).}
}
\value{
A list of the credible ranked permutations, permutations, ranked
combinations, combinations, partial hierarchies, and HDRs along with their
redundancy status.
}
\description{
`find_redundancies` compares the outputs from `get_arrangements()`,
`get_partial_hiearchies()`, and `get_ranks_by_treatment()` to identify
redundant credible hierarchies.
}
\details{
Redundancy types can be identified with numbers through 1 through 18 as follows:
\enumerate{
 \item Within ranked permutations (e.g., \eqn{(A, B)_1^2} is redundant because of \eqn{(A, B, C)_1^3}).
 \item Within permutations (e.g., \eqn{(A, B)} is redundant because of \eqn{(A, B, C)}).
 \item Within partial hierarchies (e.g., \eqn{A > B} is redundant because of \eqn{A > B > C}).
 \item Between permutations and partial hierarchies (e.g., \eqn{A > B} is redundant because of \eqn{(A, B)}).
 \item Between ranked permutations and permutations (e.g., \eqn{(A, B)} is redundant because of \eqn{(A, B)_1^2}).
 \item Between ranked combinations and combinations (e.g., \eqn{\{A, B\}} is redundant because of \eqn{\{A, B\}_1^2}).
 \item Between permutations and combinations (e.g., \eqn{\{A, B\}} is redundant because of \eqn{(A, B)}).
 \item Between ranked permutations and ranked combinations (e.g., \eqn{\{A, B\}_1^2} is redundant because of \eqn{(A, B)_1^2}).
 \item Between a ranked permutation and multiple single HDR ranks.
 \item Between a ranked combination and multiple HDRs.
 \item Within top/bottom ranked combinations.
 \item Between top/bottom single HDR ranks and a combination ranking from 1:(n_trt-1) or 2:n_trt.
 \item Between permutations ranking from 1:(n_trt-1) or 2:n_trt and a top/bottom single HDR rank.
 \item Between top/bottom single HDR ranks and a middle combination ranking 2:(n_trt-1).
 \item Between (top/bottom single HDR rank + tail-ranked permutation) and a middle ranked combination.
 \item Between tail-ranked permutations and a middle ranked combination.
 \item Between (ranked permutation + smaller ranked combination) and ranked combination.
 \item Between (single HDR rank + smaller ranked combination) and ranked combination.
}
We recommend these redundancies are identified in increasing order (1 through
18) to ensure a complete assessment and output. Otherwise, a warning will be
outputted.
}
\examples{
inputs <- prep_data(effects_matrix = dat_Thijs2008[, -1],
                    reference = "Placebo",
                    larger_better = FALSE)
algo1 <- get_arrangements(hierarchy_matrix = inputs$hierarchy_matrix,
                          threshold = 0.95)
algo2 <- get_partial_hierarchies(effects_matrix = inputs$effects_matrix,
                                 mid = 0,
                                 threshold = 0.95,
                                 larger_better = FALSE)
algo3 <- get_ranks_by_treatment(ranking_df = inputs$ranking_df,
                                threshold = 0.95,
                                print_plot = FALSE)
find_redundancies(algo1, algo2, algo3, n_trt = 5, threshold = 0.95)
}
