% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_supersets.R
\name{find_supersets}
\alias{find_supersets}
\title{Identify and optionally trim superset hierarchies}
\usage{
find_supersets(algo1, algo2, type = 1:8, trim = FALSE)
}
\arguments{
\item{algo1}{A list of data frames outputted from `consecutive()` containing
the credible hierarchies for ranked permutations, permutations, ranked
combinations, and combinations.}

\item{algo2}{A data frame outputted from `get_posets()` containing credible
partial hierarchies.}

\item{type}{A numeric vector indicating what types of supersets should be
trimmed. See details for more information. Default is 1:8.}

\item{trim}{A logical value indicating whether the output should trim the
include the identified subsets (TRUE) or not (FALSE, the default).}
}
\value{
A list of the credible ranked permutations, permutations, ranked
combinations, combinations, and partial hierarchies along with their superset
status.
}
\description{
`find_supersets` compares the outputs from `consecutive()` and `get_posets()`
to identify redundant credible hierarchies (i.e., supersets).
}
\details{
Superset types can be identified with numbers through 1 through 8 as follows:
1) Within partial hierarchies (e.g., A > B is a superset of A > B > C).
2) Between partial hierarchies and permutations (e.g., A > B is a superset of
  (A, B)).
3) Between combinations and ranked combinations (e.g., ${A, B}$ is a superset
  of ${A, B}_1^2$).
4) Between combinations and ranked combinations (e.g., ${A, B}$ is a superset
  of $(A, B)$).
5) Between ranked combinations and ranked permutations (e.g., ${A, B}_1^2$ is
  a superset of $(A, B)_1^2$).
6) Within permutations (e.g., (A, B) is a superset of (A, B, C)).
7) Between permutations and ranked permutations(e.g., $(A, B)$ is a superset
  of $(A, B)_1^2$).
8) Within ranked permutations (e.g., $(A, B)_1^2$ is a superset of
  $(A, B, C)_1^3$).

We recommend these supersets are identified in increasing order (1 through
8). Otherwise, a warning will be outputted.
}
\examples{
find_supersets(consec_outputs, posets)
}
