% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{freq}
\alias{freq}
\title{Gets frequency of columns}
\usage{
freq(data, col.range)
}
\arguments{
\item{data}{data frame}

\item{col.range}{range of columns}
}
\description{
Gets frequency of columns
}
\keyword{internal}
