% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_questions.R
\name{get_all_questions}
\alias{get_all_questions}
\title{Get all credible hierarchies}
\usage{
get_all_questions(
  inputs,
  n_trt,
  larger_better,
  thresholds,
  mid = 0,
  print_plot = FALSE,
  trim_redundant = FALSE
)
}
\arguments{
\item{inputs}{the output from `prep_data()`, which consists of a list of
`hierarchy_matrix`, `effects_matrix`, and `ranking_df`.}

\item{n_trt}{a numeric value indicating the number of treatments in the
network.}

\item{larger_better}{a logical value indicating whether larger relative
effects are better (TRUE) or not (FALSE).}

\item{thresholds}{A numeric vector containing three proportions between 0 and
1 for which a combinatorial hierarchy, partial hierarchy, and individual
ranking probabilities would be credible.}

\item{mid}{a numeric value indicating the absolute minimally important
difference. Default is 0.}

\item{print_plot}{a logical value indicating whether the rankograms should be
printed (TRUE) or not (FALSE, the default).}

\item{trim_redundant}{a logical value indicating whether the redundant
hierarchies should be trimmed from the output (TRUE) or not (FALSE, the
default).}
}
\value{
A list of data frames containing the credible hierarchies for ranked
permutations, permutations, ranked combinations, combinations, partial
hierarchies, and HDR sets.
}
\description{
`get_all_questions` runs the inputs through `get_arrangements()`,
`get_partial_hierarchies()`, and `get_all_questions()` to produce a list of
all hierarchies with empirical probabilities greater than or equal to the
threshold. These hierarchies are then run through `find_redundancies()` to
determine which of them are redundant.
}
\examples{
inputs <- prep_data(effects_matrix = dat_Thijs2008[, -1],
                    reference = "Placebo",
                    larger_better = FALSE)
get_all_questions(inputs = inputs,
                  larger_better = FALSE,
                  thresholds = c(0.9, 0.9, 0.9),
                  mid = 0,
                  print_plot = FALSE,
                  trim_redundant = FALSE)
}
