% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cred_phier.R
\name{get_cred_phier}
\alias{get_cred_phier}
\title{Get credible partial hierarchies}
\usage{
get_cred_phier(effects_matrix, MID, threshold, largerBetter)
}
\arguments{
\item{effects_matrix}{a matrix where the column headers are treatment names
and each row displays each treatment’s sampled relative effect for that
iteration.}

\item{MID}{a numeric value indicating the absolute minimally important
difference.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}

\item{largerBetter}{a logical value indicating whether larger relative
effects are better (TRUE) or not (FALSE).}
}
\value{
A data frame containing the credible partial hierarchies.
}
\description{
`get_cred_phier()` finds all partial hierarchies reflecting treatment
differences greater than or equal to `MID` and with relative frequencies
greater than or equal to the threshold.
}
\examples{
get_cred_phier(effects_matrix, 10, 0.2, TRUE)
}
