% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partial_hierarchies.R
\name{get_partial_hierarchies}
\alias{get_partial_hierarchies}
\title{Get credible partial hierarchies}
\usage{
get_partial_hierarchies(
  effects_matrix,
  mid = 0,
  threshold,
  larger_better,
  order_by = c("Size", "pi_hat")
)
}
\arguments{
\item{effects_matrix}{a matrix where the column headers are treatment names
and each row displays each treatment’s sampled relative effect for that
iteration.}

\item{mid}{a numeric value indicating the absolute minimally important
difference. Default is 0.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}

\item{larger_better}{a logical value indicating whether larger relative
effects are better (TRUE) or not (FALSE).}

\item{order_by}{a character vector consisting of "pi_hat" and "Size" only,
indicating the desired order of the arrangements within types (i.e., ranked
permutations, permutations, ranked combinations, and combinations). Default
is to order by the number of treatments ("Size") in the arrangements,
followed by the empirical probabilities ("pi_hat").}
}
\value{
A data frame containing the credible partial hierarchies.
}
\description{
`get_partial_hierarchies()` finds all partial hierarchies reflecting
treatment differences greater than or equal to `mid` and with empirical
probabilities greater than or equal to the threshold.
}
\examples{
inputs <- prep_data(effects_matrix = dat_Thijs2008[, -1],
                    reference = "Placebo",
                    larger_better = FALSE)
get_partial_hierarchies(effects_matrix = inputs$effects_matrix,
                        mid = 0,
                        threshold = 0.95,
                        larger_better = FALSE)
}
