% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_ranked_comb}
\alias{get_ranked_comb}
\title{Gets credible ranked combinations}
\usage{
get_ranked_comb(all_ranked_perm, trts)
}
\arguments{
\item{all_ranked_perm}{a data frame consisting of the ranks (`Range`) of all
observed permutations (`Var1`) and the corresponding proportion of samples
for which they were observed (`pi_hat`).}

\item{trts}{a vector of all the treatments strings.}
}
\value{
A data frame containing
  * `Range`: a string presenting the ranks corresponding to the combination
  in `Combinations`, presented as min-max.
  * `Combinations`: a string of the combination of treatments.
  * `Size`: the number of treatments in the combination.
  * `pi_hat`: the proportion of samples for which the ranked combination was
  observed.
}
\description{
get_ranked_comb()` first sorts the treatments within permutation strings to
create a combination string that ignores order. It then groups all
combinations by rank interval and sums the proportion of samples for which
they were observed (i.e., empirical probabilities).
}
\keyword{internal}
