% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_superset.R
\name{get_superset}
\alias{get_superset}
\title{Determines if credible hierarchies are supersets}
\usage{
get_superset(consec_outputs, posets, n_treatments)
}
\arguments{
\item{consec_outputs}{Output from consecutive(), which consists of a list of data frames containing the credible hierarchies for Ranked Permutations, Permutations, Ranked Combinations, and Combinations}

\item{posets}{Output from get_posets(), which consists of a data frame of the credible Posets}

\item{n_treatments}{Number of treatments}
}
\value{
A list of credible hierarchies with a column indicating which hierarchies are supersets
}
\description{
Adds an additional column to each data frame, where TRUE indicates it is a superset, and FALSE indicates it is a subset
}
\examples{
get_superset(consec_outputs,posets,3)
}
