% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{indexed_combo}
\alias{indexed_combo}
\title{Gets credible ranked combinations}
\usage{
indexed_combo(all_ranked_perm, trts)
}
\arguments{
\item{all_ranked_perm}{data frame of all ranked permutations}

\item{trts}{vector of all the treatments}
}
\description{
Gets credible ranked combinations
}
\keyword{internal}
