% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singular_treatment.R
\name{singular_treatment}
\alias{singular_treatment}
\title{Retrieves credible hierarchies for individual treatment question types}
\usage{
singular_treatment(ranking.df, threshold, printPlot = FALSE)
}
\arguments{
\item{ranking.df}{A data frame of each treatment's ranks and associated frequencies}

\item{threshold}{A number that determines which hierarchies are credible}

\item{printPlot}{Boolean value where True indicates that the rankograms will be printed (default = False)}
}
\value{
A list of data frames containing the credible hierarchies for Individual Treatments and HPD intervals
}
\description{
Finds all treatments' ranks with frequencies greater than or equal to the threshold and HPD intervals which includes the subset of ranks with the smallest possible cumulative frequency that is at least equal to the threshold
}
\examples{
singular_treatment(df,0.1,c("CBT.exp","Placebo","PCT"))
}
