% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{terminal_nonconsec}
\alias{terminal_nonconsec}
\title{Finds supersets for posets}
\usage{
terminal_nonconsec(non_consec, target)
}
\arguments{
\item{non_consec}{data frame of credible hierarchies}

\item{target}{poset we are looking for}
}
\description{
Finds supersets for posets
}
\keyword{internal}
