% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive.R
\name{consecutive}
\alias{consecutive}
\title{Retrieves credible hierarchies for consecutive (combinatorial) question types}
\usage{
consecutive(hierarchy_matrix, threshold)
}
\arguments{
\item{hierarchy_matrix}{a matrix where column headers are ranks and each row
displays the treatments assigned to each rank for that iteration.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}
}
\value{
A list of data frames containing the credible hierarchies for ranked
permutations, permutations, ranked combinations, and combinations.
}
\description{
Finds all ranked permutations, permutations, ranked combinations, and
combinations with frequencies greater than or equal to a given threshold.
}
\examples{
consecutive(hierarchy_matrix, 0.5)
}
