% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arrangements.R
\name{get_arrangements}
\alias{get_arrangements}
\title{Get credible arrangement hierarchies}
\usage{
get_arrangements(hierarchy_matrix, threshold, freq_digits = 4)
}
\arguments{
\item{hierarchy_matrix}{a matrix where column headers are ranks and each row
displays the treatments assigned to each rank for that iteration.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}

\item{freq_digits}{a numeric value indicating the desired number of decimal
places for which the relative frequencies will be rounded. Default is 4.}
}
\value{
A list of data frames containing the credible hierarchies for ranked
permutations, permutations, ranked combinations, and combinations.
}
\description{
`get_arrangements()` finds all ranked permutations, permutations, ranked
combinations, and combinations with relative frequencies greater than or
equal to a given threshold.
}
\details{
Note that the progress of the permutations compilation will be printed for
each permutation size. The user may suppress these messages using
`suppressMessages(get_arrangements())`.
}
\examples{
inputs <- prep_data(effects_matrix = dat_Thijs2008[, -1], reference = "Placebo", largerbetter = FALSE)
get_arrangements(hierarchy_matrix = inputs$hierarchy_matrix, threshold = 0.9)
}
