% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arrangements.R
\name{get_arrangements}
\alias{get_arrangements}
\title{Get credible arrangement hierarchies}
\usage{
get_arrangements(hierarchy_matrix, threshold)
}
\arguments{
\item{hierarchy_matrix}{a matrix where column headers are ranks and each row
displays the treatments assigned to each rank for that iteration.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}
}
\value{
A list of data frames containing the credible hierarchies for ranked
permutations, permutations, ranked combinations, and combinations.
}
\description{
`get_arrangements()` finds all ranked permutations, permutations, ranked
combinations, and combinations with relative frequencies greater than or
equal to a given threshold.
}
\examples{
get_arrangements(hierarchy_matrix, 0.5)
}
