% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_combo}
\alias{get_combo}
\title{Get combinations}
\usage{
get_combo(all_ranked_combo)
}
\arguments{
\item{all_ranked_combo}{a data frame consisting of the ranks (`Range`) of all
observed combinations (`Combinations`) and the corresponding proportion of
samples for which they were observed (`pi_hat`).}
}
\value{
A data frame containing
  * `Combinations`: a string of the combination of treatments.
  * `Size`: the number of treatments in the combination.
  * `pi_hat`: the proportion of samples for which the combination was
  observed  (i.e., empirical probabilities).
}
\description{
`get_combo()`groups all ranked combinations by the combination string
 (ignoring ranks), and sums the proportion of samples for which they were
 observed (i.e., empirical probabilities).
}
\keyword{internal}
