% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_combo}
\alias{get_combo}
\title{Gets credible combinations}
\usage{
get_combo(all_ranked_combo)
}
\arguments{
\item{all_ranked_combo}{data frame of all ranked combinations}
}
\description{
Gets credible combinations
}
\keyword{internal}
