% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hierarchies.R
\name{get_hierarchies}
\alias{get_hierarchies}
\title{Get all credible hierarchies}
\usage{
get_hierarchies(inputs, largerBetter, thresholds, MID = 0, printPlot = FALSE)
}
\arguments{
\item{inputs}{the output from `prep_data()`, which consists of a list of
`hierarchy_matrix`, `effects_matrix`, and `ranking_df`.}

\item{largerBetter}{a logical value indicating whether larger relative
effects are better (TRUE) or not (FALSE).}

\item{thresholds}{A numeric vector containing three proportions between 0 and
1 for which a combinatorial hierarchy, partial hierarchy, and individual
ranking probabilities would be credible.}

\item{MID}{a numeric value indicating the absolute minimally important
difference. Default is 0.}

\item{printPlot}{a logical value indicating whether the rankograms should be
printed (TRUE) or not (FALSE, the default).}
}
\value{
A list of data frames containing the credible hierarchies for ranked
permutations, permutations, ranked combinations, combinations, partial
hierarchies, individual ranking probabilities, and HPD intervals.
}
\description{
`get_hierarchies` runs the inputs through `get_arrangements()`,
`get_cred_phier()`, and `singular_treatment()` to produce a list of all
hierarchies with relative frequencies greater than or equal to the threshold.
These hierarchies are then run through `find_supersets()` to determine which
of them are supersets.
}
\examples{
get_hierarchies(inputs, "Waitlist", TRUE, c(0.5,0.6,0.7), 10)
}
