% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_perm}
\alias{get_perm}
\title{Gets credible permutations}
\usage{
get_perm(all_ranked_perm)
}
\arguments{
\item{all_ranked_perm}{data frame of all ranked permutations}
}
\description{
Gets credible permutations
}
\keyword{internal}
