% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_ranked_perm}
\alias{get_ranked_perm}
\title{Get ranked permutations}
\usage{
get_ranked_perm(hierarchy_matrix, rank_range)
}
\arguments{
\item{hierarchy_matrix}{a matrix where column headers are ranks and each row
displays the treatments assigned to each rank for that iteration.}

\item{rank_range}{a numeric vector of sequential ranks to consider.}
}
\value{
A data frame containing
  * `Range`: a string presenting the ranks corresponding to the permutation
  in `Var1`, presented as min-max.
  * `Var1`: a string of the permutation of treatments.
  * `Size`: the number of treatments in the permutation.
  * `pi_hat`: the proportion of samples for which the ranked permutation was
  observed (i.e., empirical probabilities).
}
\description{
`get_ranked_perm()` calculates the proportion of samples for which a
permutation with ranks `rank_range` was observed in `hierarchy_matrix`.
}
\keyword{internal}
