% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranks_by_treatment.R
\name{get_ranks_by_treatment}
\alias{get_ranks_by_treatment}
\title{Get credible HDR sets for each treatment}
\usage{
get_ranks_by_treatment(ranking_df, threshold, print_plot = FALSE)
}
\arguments{
\item{ranking_df}{a data frame of each treatment's ranks and associated
frequencies.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}

\item{print_plot}{a logical value indicating whether the rankograms should be
printed (TRUE) or not (FALSE, the default).}
}
\value{
A data frame containing the credible HDR set for each treatment.
}
\description{
`get_ranks_by_treatment()` finds all highest density region (HDR) sets with
empirical probabilities greater than or equal to the threshold for each
treatment. The HDR sets provide the subset of ranks with the smallest
possible cumulative empirical probability that is at least equal to the
threshold.
}
\examples{
inputs <- prep_data(effects_matrix = dat_Thijs2008[, -1], reference = "Placebo", larger_better = FALSE)
get_ranks_by_treatment(ranking_df = inputs$ranking_df, threshold = 0.9, print_plot = FALSE)
}
