% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranks_by_treatment.R
\name{get_ranks_by_treatment}
\alias{get_ranks_by_treatment}
\title{Get credible rankings and HPD sets for each treatment}
\usage{
get_ranks_by_treatment(ranking_df, threshold, print_plot = FALSE)
}
\arguments{
\item{ranking_df}{a data frame of each treatment's ranks and associated
frequencies.}

\item{threshold}{a proportion between 0 and 1 for which a hierarchy must be
observed in order to be credible.}

\item{print_plot}{a logical value indicating whether the rankograms should be
printed (TRUE) or not (FALSE, the default).}
}
\value{
A list of data frames containing the credible rankings and HPD
set for each treatment.
}
\description{
`get_ranks_by_treatment()` finds all ranks and high probability density (HPD)
sets with relative frequencies greater than or equal to the threshold for
each treatment. The HPD intervals provide the subset of ranks with the
smallest possible cumulative relative frequency that is at least equal to the
threshold.
}
\examples{
inputs <- prep_data(effects_matrix = dat_Thijs2008[, -1], reference = "Placebo", largerbetter = FALSE)
get_ranks_by_treatment(ranking_df = inputs$ranking_df, threshold = 0.9, print_plot = FALSE)
}
