% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{hpd}
\alias{hpd}
\title{Finds HPD intervals}
\usage{
hpd(ranks, threshold, freq_sum)
}
\arguments{
\item{ranks}{column of ranks and their frequency}

\item{threshold}{cutoff for what is considered credible}

\item{freq_sum}{value is always 1}
}
\description{
`hpd()`
}
\keyword{internal}
