% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_phier_redundant_within_phier}
\alias{is_phier_redundant_within_phier}
\title{Determine redundancy status of a (credible) partial hierarchy within
(credible) partial hierarchies}
\usage{
is_phier_redundant_within_phier(phier_target, larger_phier_list)
}
\arguments{
\item{phier_target}{a single character vector of treatment names (without ">"
characters) in order of a (credible) partial hierarchy to assess the
redundancy status of.}

\item{phier_larger_list}{list of data frames of larger (credible) partial
hierarchies by size, to assess the redundancy status of `phier_target`
against.}
}
\value{
`TRUE` if `phier_target` is a redundant because of any of the
  (credible) partial hierarchies listed in `larger_phier_list`.
}
\description{
`is_phier_redundant_within_phier()` determines whether a (credible) partial
hierarchy specified by `phier_target` is redundant because of any of the
(credible) partial hierarchies specified in `phier_larger_list`.
}
\keyword{internal}
