% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Data preparation}
\usage{
prep_data(df, reference, largerBetter)
}
\arguments{
\item{df}{A data frame where the column headers are treatment names and row headers are the iteration number. Each row displays each treatment’s sampled relative effect for that iteration}

\item{reference}{A string of the reference treatment's name}

\item{largerBetter}{Boolean value where True indicates larger values are better, and False otherwise}
}
\value{
\code{hierarchy.matrix} - A matrix where column headers are ranks and each row displays the treatments assigned to each rank for that iteration

\code{effects.matrix} - A matrix where the column headers are treatment names and each row displays each treatment’s sampled relative effect for that iteration

\code{ranking.df} - A data frame of each treatment's ranks and associated frequencies
}
\description{
Produces a list of inputs to be used altogether in get_hierarchies(), or individually in consecutive(), get_posets(), and singular_treatment()
}
\examples{
prep_data(df,"waitlist",TRUE)
}
