% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data to determine credible hierarchies}
\usage{
prep_data(effects_matrix, reference, largerBetter)
}
\arguments{
\item{effects_matrix}{a data frame where the column headers are treatment names
and each row displays each treatment’s sampled relative effect for that
iteration.}

\item{reference}{a character string of the reference treatment's name.}

\item{largerBetter}{a logical value indicating whether larger relative
effects are better (TRUE) or not (FALSE).}
}
\value{
\code{hierarchy_matrix} - A matrix where column headers are ranks and
each row displays the treatments assigned to each rank for that iteration.

\code{effects_matrix} - A matrix where the column headers are
treatment names and each row displays each treatment’s sampled relative
effect for that iteration.

\code{ranking_df} - A data frame of each treatment's ranks and
associated frequencies.
}
\description{
`prep_data()` produces a list of inputs to be used altogether in
`get_hierarchies()`, or individually in `get_arranagements()`,
`get_cred_phier()`, and `singular_treatment()`.
}
\examples{
prep_data(df, "waitlist", TRUE)
}
